requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
diffusers>=0.25.0
opencv-python

# Data preprocessing pipeline dependencies
reportlab>=4.0.0           # PDF generation
pdf2image>=1.16.0          # PDF to image conversion
pytesseract>=0.3.10        # OCR for reading IDs